kubernetes: share common code between services
