+++ kubernetes/service.go
@@ -1,13 +1,45 @@
 package kubernetes
 
 import (
+	"fmt"
 	"net/http"
 	"time"
 
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
+	apiv1 "k8s.io/client-go/pkg/api/v1"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/transport"
 )
+
+const (
+	defaultServicePort = 8888
+	appLabel           = "tsuru.io/app-name"
+	processLabel       = "tsuru.io/app-process"
+	swapLabel          = "tsuru.io/swapped-with"
+	appPoolLabel       = "tsuru.io/app-pool"
+	poolLabel          = "tsuru.io/pool"
+	webProcessName     = "web"
+)
+
+// ErrNoService indicates that the app has no service running
+type ErrNoService struct{ App, Process string }
+
+func (e ErrNoService) Error() string {
+	str := fmt.Sprintf("no service found for app %q", e.App)
+	if e.Process != "" {
+		str += fmt.Sprintf(" and process %q", e.Process)
+	}
+	return str
+}
+
+// ErrAppSwapped indicates when a operation cant be performed
+// because the app is swapped
+type ErrAppSwapped struct{ App, DstApp string }
+
+func (e ErrAppSwapped) Error() string {
+	return fmt.Sprintf("app %q currently swapped with %q", e.App, e.DstApp)
+}
 
 type BaseService struct {
 	Namespace   string
@@ -15,6 +47,52 @@ type BaseService struct {
 	Client      kubernetes.Interface
 	Labels      map[string]string
 	Annotations map[string]string
+}
+
+// Addresses return the addresses of every node on the same pool as the
+// app Service pool
+func (k *BaseService) Addresses(appName string) ([]string, error) {
+	service, err := k.getService(appName)
+	if err != nil {
+		return nil, err
+	}
+	var port int32
+	if len(service.Spec.Ports) > 0 {
+		port = service.Spec.Ports[0].NodePort
+	}
+	client, err := k.getClient()
+	if err != nil {
+		return nil, err
+	}
+	pool := service.Labels[appPoolLabel]
+	nodes, err := client.CoreV1().Nodes().List(metav1.ListOptions{
+		LabelSelector: fmt.Sprintf("%s=%s", poolLabel, pool),
+	})
+	if err != nil {
+		return nil, err
+	}
+	addresses := make([]string, len(nodes.Items))
+	for i := range nodes.Items {
+		addr := nodes.Items[i].Name
+		for _, a := range nodes.Items[i].Status.Addresses {
+			if a.Type == apiv1.NodeInternalIP {
+				addr = a.Address
+				break
+			}
+		}
+		addresses[i] = fmt.Sprintf("http://%s:%d", addr, port)
+	}
+	return addresses, nil
+}
+
+// Healthcheck uses the kubernetes client to check the connectivity
+func (k *BaseService) Healthcheck() error {
+	client, err := k.getClient()
+	if err != nil {
+		return err
+	}
+	_, err = client.CoreV1().Services(k.Namespace).List(metav1.ListOptions{})
+	return err
 }
 
 func (k *BaseService) getClient() (kubernetes.Interface, error) {
